refactor and clean up old flags
