+++ cypress/support/e2e.ts
@@ -33,7 +33,6 @@ beforeEach(() => {
                 'surveys-new-creation-flow': true,
                 'surveys-results-visualizations': true,
                 'auto-redirect': true,
-                'surveys-open-choice': true,
                 notebooks: true,
             })
         )
