+++ frontend/src/scenes/surveys/SurveyEdit.tsx
@@ -3,97 +3,32 @@ import './EditSurvey.scss'
 import { DndContext } from '@dnd-kit/core'
 import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable'
 import {
-    LemonBanner,
     LemonButton,
     LemonCheckbox,
     LemonCollapse,
     LemonDivider,
     LemonInput,
     LemonSelect,
-    LemonTabs,
     LemonTextArea,
     Link,
 } from '@posthog/lemon-ui'
-import clsx from 'clsx'
 import { BindLogic, useActions, useValues } from 'kea'
-import { Group } from 'kea-forms'
-import { CodeEditor } from 'lib/components/CodeEditors'
 import { FlagSelector } from 'lib/components/FlagSelector'
-import { PayGateMini } from 'lib/components/PayGateMini/PayGateMini'
-import { FEATURE_FLAGS } from 'lib/constants'
 import { Field, PureField } from 'lib/forms/Field'
-import { IconCancel, IconDelete, IconLock, IconPlus, IconPlusMini } from 'lib/lemon-ui/icons'
-import { featureFlagLogic as enabledFeaturesLogic } from 'lib/logic/featureFlagLogic'
-import React from 'react'
+import { IconCancel, IconDelete, IconLock, IconPlus } from 'lib/lemon-ui/icons'
 import { featureFlagLogic } from 'scenes/feature-flags/featureFlagLogic'
 import { FeatureFlagReleaseConditions } from 'scenes/feature-flags/FeatureFlagReleaseConditions'
 
-import {
-    AvailableFeature,
-    LinkSurveyQuestion,
-    RatingSurveyQuestion,
-    SurveyQuestion,
-    SurveyQuestionType,
-    SurveyType,
-    SurveyUrlMatchType,
-} from '~/types'
+import { LinkSurveyQuestion, RatingSurveyQuestion, SurveyQuestion, SurveyType, SurveyUrlMatchType } from '~/types'
 
-import {
-    defaultSurveyAppearance,
-    defaultSurveyFieldValues,
-    SurveyQuestionLabel,
-    SurveyUrlMatchTypeLabels,
-} from './constants'
+import { defaultSurveyAppearance, defaultSurveyFieldValues, SurveyUrlMatchTypeLabels } from './constants'
 import { SurveyAPIEditor } from './SurveyAPIEditor'
-import {
-    BaseAppearance,
-    Customization,
-    SurveyAppearance,
-    SurveyMultipleChoiceAppearance,
-    SurveyRatingAppearance,
-} from './SurveyAppearance'
-import { SurveyEditQuestionHeader } from './SurveyEditQuestionRow'
+import { Customization, SurveyAppearance } from './SurveyAppearance'
+import { HTMLEditor, PresentationTypeCard } from './SurveyAppearanceUtils'
+import { SurveyEditQuestionGroup, SurveyEditQuestionHeader } from './SurveyEditQuestionRow'
 import { SurveyFormAppearance } from './SurveyFormAppearance'
 import { SurveyEditSection, surveyLogic } from './surveyLogic'
 import { surveysLogic } from './surveysLogic'
-
-function PresentationTypeCard({
-    title,
-    description,
-    children,
-    onClick,
-    value,
-    active,
-}: {
-    title: string
-    description?: string
-    children: React.ReactNode
-    onClick: () => void
-    value: any
-    active: boolean
-}): JSX.Element {
-    return (
-        <div
-            // eslint-disable-next-line react/forbid-dom-props
-            style={{ height: 230, width: 260 }}
-            className={clsx(
-                'border rounded-md relative px-4 py-2 overflow-hidden',
-                active ? 'border-primary' : 'border-border'
-            )}
-        >
-            <p className="font-semibold m-0">{title}</p>
-            {description && <p className="m-0 text-xs">{description}</p>}
-            <div className="relative mt-2 presentation-preview">{children}</div>
-            <input
-                onClick={onClick}
-                className="opacity-0 absolute inset-0 h-full w-full cursor-pointer"
-                name="type"
-                value={value}
-                type="radio"
-            />
-        </div>
-    )
-}
 
 export default function SurveyEdit(): JSX.Element {
     const {
@@ -105,16 +40,9 @@ export default function SurveyEdit(): JSX.Element {
         selectedQuestion,
         selectedSection,
     } = useValues(surveyLogic)
-    const {
-        setSurveyValue,
-        setDefaultForQuestionType,
-        setWritingHTMLDescription,
-        resetTargeting,
-        setSelectedQuestion,
-        setSelectedSection,
-    } = useActions(surveyLogic)
+    const { setSurveyValue, setWritingHTMLDescription, resetTargeting, setSelectedQuestion, setSelectedSection } =
+        useActions(surveyLogic)
     const { surveysMultipleQuestionsAvailable } = useValues(surveysLogic)
-    const { featureFlags } = useValues(enabledFeaturesLogic)
     const sortedItemIds = survey.questions.map((_, idx) => idx.toString())
 
     function onSortEnd({ oldIndex, newIndex }: { oldIndex: number; newIndex: number }): void {
@@ -189,514 +117,11 @@ export default function SurveyEdit(): JSX.Element {
                                                                 />
                                                             ),
                                                             content: (
-                                                                <Group name={`questions.${index}`} key={index}>
-                                                                    <div className="flex flex-col gap-2">
-                                                                        <Field
-                                                                            name="type"
-                                                                            label="Question type"
-                                                                            className="max-w-60"
-                                                                        >
-                                                                            <LemonSelect
-                                                                                data-attr={`survey-question-type-${index}`}
-                                                                                onSelect={(newType) => {
-                                                                                    const isEditingQuestion =
-                                                                                        defaultSurveyFieldValues[
-                                                                                            question.type
-                                                                                        ].questions[0].question !==
-                                                                                        question.question
-                                                                                    const isEditingDescription =
-                                                                                        defaultSurveyFieldValues[
-                                                                                            question.type
-                                                                                        ].questions[0].description !==
-                                                                                        question.description
-                                                                                    const isEditingThankYouMessage =
-                                                                                        defaultSurveyFieldValues[
-                                                                                            question.type
-                                                                                        ].appearance
-                                                                                            .thankYouMessageHeader !==
-                                                                                        survey.appearance
-                                                                                            .thankYouMessageHeader
-                                                                                    setDefaultForQuestionType(
-                                                                                        index,
-                                                                                        newType,
-                                                                                        isEditingQuestion,
-                                                                                        isEditingDescription,
-                                                                                        isEditingThankYouMessage
-                                                                                    )
-                                                                                }}
-                                                                                options={[
-                                                                                    {
-                                                                                        label: SurveyQuestionLabel[
-                                                                                            SurveyQuestionType.Open
-                                                                                        ],
-                                                                                        value: SurveyQuestionType.Open,
-                                                                                        tooltip: () => (
-                                                                                            <BaseAppearance
-                                                                                                preview
-                                                                                                onSubmit={() =>
-                                                                                                    undefined
-                                                                                                }
-                                                                                                appearance={{
-                                                                                                    ...survey.appearance,
-                                                                                                    whiteLabel: true,
-                                                                                                }}
-                                                                                                question={{
-                                                                                                    type: SurveyQuestionType.Open,
-                                                                                                    question:
-                                                                                                        'Share your thoughts',
-                                                                                                    description:
-                                                                                                        'Optional form description',
-                                                                                                }}
-                                                                                            />
-                                                                                        ),
-                                                                                    },
-                                                                                    {
-                                                                                        label: 'Link',
-                                                                                        value: SurveyQuestionType.Link,
-                                                                                        tooltip: () => (
-                                                                                            <BaseAppearance
-                                                                                                preview
-                                                                                                onSubmit={() =>
-                                                                                                    undefined
-                                                                                                }
-                                                                                                appearance={{
-                                                                                                    ...survey.appearance,
-                                                                                                    whiteLabel: true,
-                                                                                                }}
-                                                                                                question={{
-                                                                                                    type: SurveyQuestionType.Link,
-                                                                                                    question:
-                                                                                                        'Do you want to join our upcoming webinar?',
-                                                                                                    buttonText:
-                                                                                                        'Register',
-                                                                                                    link: '',
-                                                                                                }}
-                                                                                            />
-                                                                                        ),
-                                                                                    },
-                                                                                    {
-                                                                                        label: 'Rating',
-                                                                                        value: SurveyQuestionType.Rating,
-                                                                                        tooltip: () => (
-                                                                                            <SurveyRatingAppearance
-                                                                                                preview
-                                                                                                onSubmit={() =>
-                                                                                                    undefined
-                                                                                                }
-                                                                                                appearance={{
-                                                                                                    ...survey.appearance,
-                                                                                                    whiteLabel: true,
-                                                                                                }}
-                                                                                                ratingSurveyQuestion={{
-                                                                                                    question:
-                                                                                                        'How satisfied are you with our product?',
-                                                                                                    description:
-                                                                                                        'Optional form description.',
-                                                                                                    display: 'number',
-                                                                                                    lowerBoundLabel:
-                                                                                                        'Not great',
-                                                                                                    upperBoundLabel:
-                                                                                                        'Fantastic',
-                                                                                                    scale: 5,
-                                                                                                    type: SurveyQuestionType.Rating,
-                                                                                                }}
-                                                                                            />
-                                                                                        ),
-                                                                                    },
-                                                                                    ...[
-                                                                                        {
-                                                                                            label: 'Single choice select',
-                                                                                            value: SurveyQuestionType.SingleChoice,
-                                                                                            tooltip: () => (
-                                                                                                <SurveyMultipleChoiceAppearance
-                                                                                                    initialChecked={[0]}
-                                                                                                    preview
-                                                                                                    onSubmit={() =>
-                                                                                                        undefined
-                                                                                                    }
-                                                                                                    appearance={{
-                                                                                                        ...survey.appearance,
-                                                                                                        whiteLabel:
-                                                                                                            true,
-                                                                                                    }}
-                                                                                                    multipleChoiceQuestion={{
-                                                                                                        type: SurveyQuestionType.SingleChoice,
-                                                                                                        choices: [
-                                                                                                            'Yes',
-                                                                                                            'No',
-                                                                                                        ],
-                                                                                                        question:
-                                                                                                            'Have you found this tutorial useful?',
-                                                                                                    }}
-                                                                                                />
-                                                                                            ),
-                                                                                        },
-                                                                                        {
-                                                                                            label: 'Multiple choice select',
-                                                                                            value: SurveyQuestionType.MultipleChoice,
-                                                                                            tooltip: () => (
-                                                                                                <SurveyMultipleChoiceAppearance
-                                                                                                    initialChecked={[
-                                                                                                        0, 1,
-                                                                                                    ]}
-                                                                                                    preview
-                                                                                                    onSubmit={() =>
-                                                                                                        undefined
-                                                                                                    }
-                                                                                                    appearance={{
-                                                                                                        ...survey.appearance,
-                                                                                                        whiteLabel:
-                                                                                                            true,
-                                                                                                    }}
-                                                                                                    multipleChoiceQuestion={{
-                                                                                                        type: SurveyQuestionType.MultipleChoice,
-                                                                                                        choices: [
-                                                                                                            'Tutorials',
-                                                                                                            'Customer case studies',
-                                                                                                            'Product announcements',
-                                                                                                        ],
-                                                                                                        question:
-                                                                                                            'Which types of content would you like to see more of?',
-                                                                                                    }}
-                                                                                                />
-                                                                                            ),
-                                                                                        },
-                                                                                    ],
-                                                                                ]}
-                                                                            />
-                                                                        </Field>
-                                                                        <Field name="question" label="Label">
-                                                                            <LemonInput value={question.question} />
-                                                                        </Field>
-
-                                                                        <Field
-                                                                            name="description"
-                                                                            label="Description (optional)"
-                                                                        >
-                                                                            {({ value, onChange }) => (
-                                                                                <HTMLEditor
-                                                                                    value={value}
-                                                                                    onChange={onChange}
-                                                                                    writingHTMLDescription={
-                                                                                        writingHTMLDescription
-                                                                                    }
-                                                                                    setWritingHTMLDescription={
-                                                                                        setWritingHTMLDescription
-                                                                                    }
-                                                                                />
-                                                                            )}
-                                                                        </Field>
-                                                                        {survey.questions.length > 1 && (
-                                                                            <Field name="optional" className="my-2">
-                                                                                <LemonCheckbox
-                                                                                    label="Optional"
-                                                                                    checked={!!question.optional}
-                                                                                />
-                                                                            </Field>
-                                                                        )}
-                                                                        {question.type === SurveyQuestionType.Link && (
-                                                                            <Field
-                                                                                name="link"
-                                                                                label="Link"
-                                                                                info="Make sure to include https:// in the url."
-                                                                            >
-                                                                                <LemonInput
-                                                                                    value={question.link || ''}
-                                                                                    placeholder="https://posthog.com"
-                                                                                />
-                                                                            </Field>
-                                                                        )}
-                                                                        {question.type ===
-                                                                            SurveyQuestionType.Rating && (
-                                                                            <div className="flex flex-col gap-2">
-                                                                                <div className="flex flex-row gap-4">
-                                                                                    <Field
-                                                                                        name="display"
-                                                                                        label="Display type"
-                                                                                        className="w-1/2"
-                                                                                    >
-                                                                                        <LemonSelect
-                                                                                            options={[
-                                                                                                {
-                                                                                                    label: 'Number',
-                                                                                                    value: 'number',
-                                                                                                },
-                                                                                                {
-                                                                                                    label: 'Emoji',
-                                                                                                    value: 'emoji',
-                                                                                                },
-                                                                                            ]}
-                                                                                        />
-                                                                                    </Field>
-                                                                                    <Field
-                                                                                        name="scale"
-                                                                                        label="Scale"
-                                                                                        className="w-1/2"
-                                                                                    >
-                                                                                        <LemonSelect
-                                                                                            options={[
-                                                                                                ...(question.display ===
-                                                                                                'emoji'
-                                                                                                    ? [
-                                                                                                          {
-                                                                                                              label: '1 - 3',
-                                                                                                              value: 3,
-                                                                                                          },
-                                                                                                      ]
-                                                                                                    : []),
-                                                                                                {
-                                                                                                    label: '1 - 5',
-                                                                                                    value: 5,
-                                                                                                },
-                                                                                                ...(question.display ===
-                                                                                                'number'
-                                                                                                    ? [
-                                                                                                          {
-                                                                                                              label: '0 - 10',
-                                                                                                              value: 10,
-                                                                                                          },
-                                                                                                      ]
-                                                                                                    : []),
-                                                                                            ]}
-                                                                                        />
-                                                                                    </Field>
-                                                                                </div>
-                                                                                <div className="flex flex-row gap-4">
-                                                                                    <Field
-                                                                                        name="lowerBoundLabel"
-                                                                                        label="Lower bound label"
-                                                                                        className="w-1/2"
-                                                                                    >
-                                                                                        <LemonInput
-                                                                                            value={
-                                                                                                question.lowerBoundLabel ||
-                                                                                                ''
-                                                                                            }
-                                                                                        />
-                                                                                    </Field>
-                                                                                    <Field
-                                                                                        name="upperBoundLabel"
-                                                                                        label="Upper bound label"
-                                                                                        className="w-1/2"
-                                                                                    >
-                                                                                        <LemonInput
-                                                                                            value={
-                                                                                                question.upperBoundLabel ||
-                                                                                                ''
-                                                                                            }
-                                                                                        />
-                                                                                    </Field>
-                                                                                </div>
-                                                                            </div>
-                                                                        )}
-                                                                        {(question.type ===
-                                                                            SurveyQuestionType.SingleChoice ||
-                                                                            question.type ===
-                                                                                SurveyQuestionType.MultipleChoice) && (
-                                                                            <div className="flex flex-col gap-2">
-                                                                                <Field name="hasOpenChoice">
-                                                                                    {({
-                                                                                        value: hasOpenChoice,
-                                                                                        onChange: toggleHasOpenChoice,
-                                                                                    }) => (
-                                                                                        <Field
-                                                                                            name="choices"
-                                                                                            label="Choices"
-                                                                                        >
-                                                                                            {({ value, onChange }) => (
-                                                                                                <div className="flex flex-col gap-2">
-                                                                                                    {(value || []).map(
-                                                                                                        (
-                                                                                                            choice: string,
-                                                                                                            index: number
-                                                                                                        ) => {
-                                                                                                            const isOpenChoice =
-                                                                                                                hasOpenChoice &&
-                                                                                                                index ===
-                                                                                                                    value?.length -
-                                                                                                                        1
-                                                                                                            return (
-                                                                                                                <div
-                                                                                                                    className="flex flex-row gap-2 relative"
-                                                                                                                    key={
-                                                                                                                        index
-                                                                                                                    }
-                                                                                                                >
-                                                                                                                    <LemonInput
-                                                                                                                        value={
-                                                                                                                            choice
-                                                                                                                        }
-                                                                                                                        fullWidth
-                                                                                                                        onChange={(
-                                                                                                                            val
-                                                                                                                        ) => {
-                                                                                                                            const newChoices =
-                                                                                                                                [
-                                                                                                                                    ...value,
-                                                                                                                                ]
-                                                                                                                            newChoices[
-                                                                                                                                index
-                                                                                                                            ] =
-                                                                                                                                val
-                                                                                                                            onChange(
-                                                                                                                                newChoices
-                                                                                                                            )
-                                                                                                                        }}
-                                                                                                                    />
-                                                                                                                    {isOpenChoice && (
-                                                                                                                        <span className="question-choice-open-ended-footer">
-                                                                                                                            open-ended
-                                                                                                                        </span>
-                                                                                                                    )}
-                                                                                                                    <LemonButton
-                                                                                                                        icon={
-                                                                                                                            <IconDelete />
-                                                                                                                        }
-                                                                                                                        size="small"
-                                                                                                                        status="muted"
-                                                                                                                        noPadding
-                                                                                                                        onClick={() => {
-                                                                                                                            const newChoices =
-                                                                                                                                [
-                                                                                                                                    ...value,
-                                                                                                                                ]
-                                                                                                                            newChoices.splice(
-                                                                                                                                index,
-                                                                                                                                1
-                                                                                                                            )
-                                                                                                                            onChange(
-                                                                                                                                newChoices
-                                                                                                                            )
-                                                                                                                            if (
-                                                                                                                                isOpenChoice
-                                                                                                                            ) {
-                                                                                                                                toggleHasOpenChoice(
-                                                                                                                                    false
-                                                                                                                                )
-                                                                                                                            }
-                                                                                                                        }}
-                                                                                                                    />
-                                                                                                                </div>
-                                                                                                            )
-                                                                                                        }
-                                                                                                    )}
-                                                                                                    <div className="w-fit flex flex-row flex-wrap gap-2">
-                                                                                                        {(value || [])
-                                                                                                            .length <
-                                                                                                            6 && (
-                                                                                                            <>
-                                                                                                                <LemonButton
-                                                                                                                    icon={
-                                                                                                                        <IconPlusMini />
-                                                                                                                    }
-                                                                                                                    type="secondary"
-                                                                                                                    fullWidth={
-                                                                                                                        false
-                                                                                                                    }
-                                                                                                                    onClick={() => {
-                                                                                                                        if (
-                                                                                                                            !value
-                                                                                                                        ) {
-                                                                                                                            onChange(
-                                                                                                                                [
-                                                                                                                                    '',
-                                                                                                                                ]
-                                                                                                                            )
-                                                                                                                        } else if (
-                                                                                                                            hasOpenChoice
-                                                                                                                        ) {
-                                                                                                                            const newChoices =
-                                                                                                                                value.slice(
-                                                                                                                                    0,
-                                                                                                                                    -1
-                                                                                                                                )
-                                                                                                                            newChoices.push(
-                                                                                                                                ''
-                                                                                                                            )
-                                                                                                                            newChoices.push(
-                                                                                                                                value[
-                                                                                                                                    value.length -
-                                                                                                                                        1
-                                                                                                                                ]
-                                                                                                                            )
-                                                                                                                            onChange(
-                                                                                                                                newChoices
-                                                                                                                            )
-                                                                                                                        } else {
-                                                                                                                            onChange(
-                                                                                                                                [
-                                                                                                                                    ...value,
-                                                                                                                                    '',
-                                                                                                                                ]
-                                                                                                                            )
-                                                                                                                        }
-                                                                                                                    }}
-                                                                                                                >
-                                                                                                                    Add
-                                                                                                                    choice
-                                                                                                                </LemonButton>
-                                                                                                                {featureFlags[
-                                                                                                                    FEATURE_FLAGS
-                                                                                                                        .SURVEYS_OPEN_CHOICE
-                                                                                                                ] &&
-                                                                                                                    !hasOpenChoice && (
-                                                                                                                        <LemonButton
-                                                                                                                            icon={
-                                                                                                                                <IconPlusMini />
-                                                                                                                            }
-                                                                                                                            type="secondary"
-                                                                                                                            fullWidth={
-                                                                                                                                false
-                                                                                                                            }
-                                                                                                                            onClick={() => {
-                                                                                                                                if (
-                                                                                                                                    !value
-                                                                                                                                ) {
-                                                                                                                                    onChange(
-                                                                                                                                        [
-                                                                                                                                            'Other',
-                                                                                                                                        ]
-                                                                                                                                    )
-                                                                                                                                } else {
-                                                                                                                                    onChange(
-                                                                                                                                        [
-                                                                                                                                            ...value,
-                                                                                                                                            'Other',
-                                                                                                                                        ]
-                                                                                                                                    )
-                                                                                                                                }
-                                                                                                                                toggleHasOpenChoice(
-                                                                                                                                    true
-                                                                                                                                )
-                                                                                                                            }}
-                                                                                                                        >
-                                                                                                                            Add
-                                                                                                                            open-ended
-                                                                                                                            choice
-                                                                                                                        </LemonButton>
-                                                                                                                    )}
-                                                                                                            </>
-                                                                                                        )}
-                                                                                                    </div>
-                                                                                                </div>
-                                                                                            )}
-                                                                                        </Field>
-                                                                                    )}
-                                                                                </Field>
-                                                                            </div>
-                                                                        )}
-                                                                        <Field name="buttonText" label="Button text">
-                                                                            <LemonInput
-                                                                                value={
-                                                                                    question.buttonText === undefined
-                                                                                        ? survey.appearance
-                                                                                              .submitButtonText
-                                                                                        : question.buttonText
-                                                                                }
-                                                                            />
-                                                                        </Field>
-                                                                    </div>
-                                                                </Group>
+                                                                <SurveyEditQuestionGroup
+                                                                    index={index}
+                                                                    key={index}
+                                                                    question={question}
+                                                                />
                                                             ),
                                                         })
                                                     ),
@@ -791,39 +216,37 @@ export default function SurveyEdit(): JSX.Element {
                                         </SortableContext>
                                     </DndContext>
                                     <div className="flex gap-2">
-                                        {featureFlags[FEATURE_FLAGS.SURVEYS_MULTIPLE_QUESTIONS] && (
-                                            <div className="flex items-center gap-2 mt-2">
-                                                <LemonButton
-                                                    type="secondary"
-                                                    className="w-max"
-                                                    icon={<IconPlus />}
-                                                    sideIcon={
-                                                        surveysMultipleQuestionsAvailable ? null : (
-                                                            <IconLock className="ml-1 text-base text-muted" />
-                                                        )
-                                                    }
-                                                    disabledReason={
-                                                        surveysMultipleQuestionsAvailable
-                                                            ? null
-                                                            : 'Subscribe to surveys for multiple questions'
-                                                    }
-                                                    onClick={() => {
-                                                        setSurveyValue('questions', [
-                                                            ...survey.questions,
-                                                            { ...defaultSurveyFieldValues.open.questions[0] },
-                                                        ])
-                                                        setSelectedQuestion(survey.questions.length)
-                                                    }}
-                                                >
-                                                    Add question
-                                                </LemonButton>
-                                                {!surveysMultipleQuestionsAvailable && (
-                                                    <Link to={'/organization/billing'} target="_blank" targetBlankIcon>
-                                                        Subscribe
-                                                    </Link>
-                                                )}
-                                            </div>
-                                        )}
+                                        <div className="flex items-center gap-2 mt-2">
+                                            <LemonButton
+                                                type="secondary"
+                                                className="w-max"
+                                                icon={<IconPlus />}
+                                                sideIcon={
+                                                    surveysMultipleQuestionsAvailable ? null : (
+                                                        <IconLock className="ml-1 text-base text-muted" />
+                                                    )
+                                                }
+                                                disabledReason={
+                                                    surveysMultipleQuestionsAvailable
+                                                        ? null
+                                                        : 'Subscribe to surveys for multiple questions'
+                                                }
+                                                onClick={() => {
+                                                    setSurveyValue('questions', [
+                                                        ...survey.questions,
+                                                        { ...defaultSurveyFieldValues.open.questions[0] },
+                                                    ])
+                                                    setSelectedQuestion(survey.questions.length)
+                                                }}
+                                            >
+                                                Add question
+                                            </LemonButton>
+                                            {!surveysMultipleQuestionsAvailable && (
+                                                <Link to={'/organization/billing'} target="_blank" targetBlankIcon>
+                                                    Subscribe
+                                                </Link>
+                                            )}
+                                        </div>
                                         {!survey.appearance.displayThankYouMessage && (
                                             <LemonButton
                                                 type="secondary"
@@ -1119,111 +542,5 @@ export default function SurveyEdit(): JSX.Element {
                 />
             </div>
         </div>
-    )
-}
-
-export function HTMLEditor({
-    value,
-    onChange,
-    writingHTMLDescription,
-    setWritingHTMLDescription,
-    textPlaceholder,
-}: {
-    value?: string
-    onChange: (value: any) => void
-    writingHTMLDescription: boolean
-    setWritingHTMLDescription: (writingHTML: boolean) => void
-    textPlaceholder?: string
-}): JSX.Element {
-    const { surveysHTMLAvailable } = useValues(surveysLogic)
-    return (
-        <>
-            <LemonTabs
-                activeKey={writingHTMLDescription ? 'html' : 'text'}
-                onChange={(key) => setWritingHTMLDescription(key === 'html')}
-                tabs={[
-                    {
-                        key: 'text',
-                        label: <span className="text-sm">Text</span>,
-                        content: (
-                            <LemonTextArea
-                                minRows={2}
-                                value={value}
-                                onChange={(v) => onChange(v)}
-                                placeholder={textPlaceholder}
-                            />
-                        ),
-                    },
-                    {
-                        key: 'html',
-                        label: (
-                            <div>
-                                <span className="text-sm">HTML</span>
-                                {!surveysHTMLAvailable && <IconLock className="ml-2" />}
-                            </div>
-                        ),
-                        content: (
-                            <div>
-                                {surveysHTMLAvailable ? (
-                                    <CodeEditor
-                                        className="border"
-                                        language="html"
-                                        value={value}
-                                        onChange={(v) => onChange(v ?? '')}
-                                        height={150}
-                                        options={{
-                                            minimap: {
-                                                enabled: false,
-                                            },
-                                            scrollbar: {
-                                                alwaysConsumeMouseWheel: false,
-                                            },
-                                            wordWrap: 'on',
-                                            scrollBeyondLastLine: false,
-                                            automaticLayout: true,
-                                            fixedOverflowWidgets: true,
-                                            lineNumbers: 'off',
-                                            glyphMargin: false,
-                                            folding: false,
-                                        }}
-                                    />
-                                ) : (
-                                    <PayGateMini feature={AvailableFeature.SURVEYS_TEXT_HTML}>
-                                        <CodeEditor
-                                            className="border"
-                                            language="html"
-                                            value={value}
-                                            onChange={(v) => onChange(v ?? '')}
-                                            height={150}
-                                            options={{
-                                                minimap: {
-                                                    enabled: false,
-                                                },
-                                                scrollbar: {
-                                                    alwaysConsumeMouseWheel: false,
-                                                },
-                                                wordWrap: 'on',
-                                                scrollBeyondLastLine: false,
-                                                automaticLayout: true,
-                                                fixedOverflowWidgets: true,
-                                                lineNumbers: 'off',
-                                                glyphMargin: false,
-                                                folding: false,
-                                            }}
-                                        />
-                                    </PayGateMini>
-                                )}
-                            </div>
-                        ),
-                    },
-                ]}
-            />
-            {value && value?.toLowerCase().includes('<script') && (
-                <LemonBanner type="warning">
-                    Scripts won't run in the survey popover and we'll remove these on save. Use the API question mode to
-                    run your own scripts in surveys.
-                </LemonBanner>
-            )}
-        </>
     )
 }
